- Refactored conversion code
