+++ src/transform/convert/generate1xxFields.js
@@ -4,7 +4,7 @@
 *
 * Publication archives record transformer for the Melinda record batch import system
 *
-* Copyright (C) 2019-2020 University Of Helsinki (The National Library Of Finland)
+* Copyright (C) 2019-2021 University Of Helsinki (The National Library Of Finland)
 *
 * This file is part of melinda-record-import-transformer-publication-archives
 *
@@ -26,12 +26,34 @@
 *
 */
 
+export function generate100and700({getFieldValues}) {
+  const writers = generateWriters();
+  const editors = generateEditors();
 
-import {xmlToObject} from './utils';
+  return writers.concat(editors);
 
-run();
+  function generateWriters() {
+    const values = getFieldValues(p => [
+      'dc.contributor.author',
+      'dc.creator'
+    ].includes(p));
+    return values.map((v, index) => ({
+      tag: index === 0 ? '100' : '700', ind1: '1', ind2: '',
+      subfields: [
+        {code: 'a', value: `${v},`},
+        {code: 'e', value: 'kirjoittaja.'}
+      ]
+    }));
+  }
 
-async function run() {
-  const {'OAI-PMH': {GetRecord}} = await xmlToObject(process.stdin);
-  console.log(JSON.stringify(GetRecord[0].record, undefined, 2)); // eslint-disable-line no-console
+  function generateEditors() {
+    const values = getFieldValues('dc.contributor.editor');
+    return values.map(v => ({
+      tag: '700', ind1: '1', ind2: '',
+      subfields: [
+        {code: 'a', value: `${v},`},
+        {code: 'e', value: 'toimittaja.'}
+      ]
+    }));
+  }
 }
