+++ src/transform/convert/generateSidFields.js
@@ -4,7 +4,7 @@
 *
 * Publication archives record transformer for the Melinda record batch import system
 *
-* Copyright (C) 2019-2020 University Of Helsinki (The National Library Of Finland)
+* Copyright (C) 2019-2021 University Of Helsinki (The National Library Of Finland)
 *
 * This file is part of melinda-record-import-transformer-publication-archives
 *
@@ -26,12 +26,27 @@
 *
 */
 
+export function generateSID({getFieldValues}, sourceMap) {
+  const values = getFieldValues('dc.identifier.uri');
+  const baseUrlRegex = /https?:\/\/(?<source>[^?#/]+)/u;
+  const handleRegex = /(?<handle>\/[^/]+\/[^/]+$)/u;
 
-import {xmlToObject} from './utils';
+  const validSidValues = values.reduce((acc, hdl) => {
+    const {source} = hdl.match(baseUrlRegex).groups;
+    const {handle} = hdl.match(handleRegex).groups;
 
-run();
+    if (source !== null && handle !== null && Object.prototype.hasOwnProperty.call(sourceMap, source)) {
+      return acc.concat({source: sourceMap[source], handle});
+    }
 
-async function run() {
-  const {'OAI-PMH': {GetRecord}} = await xmlToObject(process.stdin);
-  console.log(JSON.stringify(GetRecord[0].record, undefined, 2)); // eslint-disable-line no-console
+    return acc;
+  }, []);
+
+  return validSidValues.length > 0 ? validSidValues.map(v => (
+    {tag: 'SID', ind1: '', ind2: '',
+      subfields: [
+        {code: 'c', value: v.handle},
+        {code: 'b', value: v.source}
+      ]}
+  )) : [];
 }
