+++ src/transform/filter/filterByFileType.js
@@ -4,7 +4,7 @@
 *
 * Publication archives record transformer for the Melinda record batch import system
 *
-* Copyright (C) 2019-2020 University Of Helsinki (The National Library Of Finland)
+* Copyright (C) 2019-2021 University Of Helsinki (The National Library Of Finland)
 *
 * This file is part of melinda-record-import-transformer-publication-archives
 *
@@ -26,12 +26,14 @@
 *
 */
 
+import {Error as NotSupportedError} from '@natlibfi/melinda-commons';
+import {getFileTypesInformation} from '../common';
 
-import {xmlToObject} from './utils';
+export function filterByFileType(record) {
+  const values = getFileTypesInformation(record) || false;
+  if (!values || values.length === 0) {
+    throw new NotSupportedError(null, null, 'Conversion without filetype specification is not supported');
+  }
 
-run();
-
-async function run() {
-  const {'OAI-PMH': {GetRecord}} = await xmlToObject(process.stdin);
-  console.log(JSON.stringify(GetRecord[0].record, undefined, 2)); // eslint-disable-line no-console
+  return values;
 }
