+++ src/transform/filter/index.js
@@ -4,7 +4,7 @@
 *
 * Publication archives record transformer for the Melinda record batch import system
 *
-* Copyright (C) 2019-2020 University Of Helsinki (The National Library Of Finland)
+* Copyright (C) 2019-2021 University Of Helsinki (The National Library Of Finland)
 *
 * This file is part of melinda-record-import-transformer-publication-archives
 *
@@ -26,12 +26,21 @@
 *
 */
 
+import {getInputFields, createValueInterface} from '../common';
+import {filterByFileType} from './filterByFileType';
+import {filterByMaterialType} from './filterByMaterialType';
 
-import {xmlToObject} from './utils';
+export default (record) => {
+  const inputFields = getInputFields(record);
+  const fieldValueInterface = createValueInterface(inputFields);
 
-run();
+  const recordFilters = {
+    raw: [filterByFileType],
+    interface: [filterByMaterialType]
+  };
 
-async function run() {
-  const {'OAI-PMH': {GetRecord}} = await xmlToObject(process.stdin);
-  console.log(JSON.stringify(GetRecord[0].record, undefined, 2)); // eslint-disable-line no-console
-}
+  recordFilters.raw.forEach(f => f(record));
+  recordFilters.interface.forEach(f => f(fieldValueInterface));
+
+  return fieldValueInterface;
+};
