+++ src/transform/utils.js
@@ -26,12 +26,32 @@
 *
 */
 
+import {Parser} from 'xml2js';
 
-import {xmlToObject} from './utils';
+export async function xmlToObject(stream) {
+  const str = await readToString();
+  return toObject();
 
-run();
+  function readToString() {
+    return new Promise((resolve, reject) => {
+      const list = [];
 
-async function run() {
-  const {'OAI-PMH': {GetRecord}} = await xmlToObject(process.stdin);
-  console.log(JSON.stringify(GetRecord[0].record, undefined, 2)); // eslint-disable-line no-console
+      stream
+        .on('error', reject)
+        .on('data', chunk => list.push(chunk)) // eslint-disable-line functional/immutable-data
+        .on('end', () => resolve(list.join('')));
+    });
+  }
+
+  function toObject() {
+    return new Promise((resolve, reject) => {
+      new Parser().parseString(str, (err, obj) => {
+        if (err) {
+          return reject(err);
+        }
+
+        resolve(obj);
+      });
+    });
+  }
 }
